feat(libmc): NBT 新增 Nbt_SetFloat，Entity_SetHealth 改用 float
- Nbt.h 新增 Nbt_SetFloat：经 McfStrRef_FromFloat 转换后写入 nbt heap
- Entity_SetHealth 参数 int → float，改调 Nbt_SetFloat（生命值本为浮点）
- 重新生成 _ll_libmc.mch；init_vm.mcfunction 清理注释

+++ src/resources/libmc/include/entity/Entity.h
@@ -92,7 +92,7 @@ Entity_GetEyeHeight(Entity entity)
     return entity ? entity->eyeHeight : 0.0f;
 }
 
-void Entity_SetHealth(Target entity, int health);
+void Entity_SetHealth(Target entity, float health);
 
 void Entity_Kill(Target target);
